updated unit test cases
