--- packages/camera/src/__mocks__/expo-camera.js
@@ -1,3 +1,0 @@
-module.exports = () => ({
-  isAvailableAsync: () => true,
-});
