--- packages/camera/src/__mocks__/i18n.js
@@ -1,8 +1,0 @@
-module.exports = () => ({
-  useTranslation: () => ({
-    t: (str) => str,
-    i18n: {
-      changeLanguage: () => Promise.resolve(),
-    },
-  }),
-});
