+++ packages/camera/src/components/Controls/__tests__/FullScreenButton.test.js
@@ -1,10 +1,14 @@
 import React from 'react';
-import { render, screen } from '../../../testUtils/i18Test-utils';
+import { render, screen } from '../../../testUtils/wrapper-utils';
 import FullScreenButton from '../FullScreenButton';
 
 describe('Fullscreen Button component', () => {
-  it('renders the correct label', () => {
-    render(<FullScreenButton />, { withI18: true });
-    expect(screen.getByText('Fullscreen')).toBeOnTheScreen();
+  it('renders the fullscreen-exit label', () => {
+    render(<FullScreenButton isInFullScreenMode />);
+    expect(screen.getByText('fullscreen-exit')).toBeOnTheScreen();
+  });
+  it('renders the fullscreen label', () => {
+    render(<FullScreenButton isInFullScreenMode={false} />);
+    expect(screen.getByText('fullscreen')).toBeOnTheScreen();
   });
 });
