+++ packages/camera/src/setUpTests.js
@@ -1,4 +1,20 @@
 import { jest } from '@jest/globals';
 import '@testing-library/jest-dom';
 
-jest.mock('./__mocks__/expo-camera');
+jest.mock('expo-camera', () => ({
+  isAvailableAsync: () => true,
+}));
+
+jest.mock('@expo/vector-icons', () => ({
+  MaterialCommunityIcons: ({ name }) => name,
+}));
+
+jest.mock('react-i18next', () => ({
+  initReactI18next: { type: '3rdParty', init: () => {} },
+  useTranslation: () => ({
+    t: (str) => str,
+    i18n: {
+      changeLanguage: () => Promise.resolve(),
+    },
+  }),
+}));
